requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
requests
numpy
sqlalchemy
python-dotenv
pydantic